Log in database integration.
Progress:
Logging in now queries the database and checks whether the users exist using that.
Additionally registering as a new user will write the credentials into the database.

Current Issues :
The login system works however it is not making use of the encryption system.

The MFA system is broken because the login page is not storing the email in the sessionStorage correctly.